% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{build.A.matrix}
\alias{build.A.matrix}
\title{Observation matrix computation.}
\usage{
build.A.matrix(operator_list, locs, i = NULL)
}
\arguments{
\item{operator_list}{A list for the operator created using \code{"create_operator"}.}

\item{locs}{A list of measurement locations.}

\item{i}{A numerical value for the index in \code{"locs"} for
which the observation matrix should be computed.}
}
\value{
An observation matrix.
}
\description{
Computes observation matrix linking basis functions to
   measurement locations.
}
\details{
This function is supplementary and internally used.
}
\examples{
  \dontrun{
  build.A.matrix(...)
  }

}
\seealso{
\code{\link{estimateLong}}
}
