% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charfuncUtil.R
\name{calc_moment_density}
\alias{calc_moment_density}
\title{takes output from characteristic_function_to_density and computes
mean, variance, skewness, kurtosis}
\usage{
calc_moment_density(dens, central = F)
}
\arguments{
\item{dens}{- (list) output from characteristic_function_to_density}
}
\value{
vec - (4x1) expectiation, variance, skewness, kurtosis
}
\description{
takes output from characteristic_function_to_density and computes
mean, variance, skewness, kurtosis
}
