% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charfuncUtil.R
\name{calc_moment_density_2d}
\alias{calc_moment_density_2d}
\title{takes output from characteristic_function_to_density and computes
mean, variance, skewness, kurtosis}
\usage{
calc_moment_density_2d(dens, central = F)
}
\arguments{
\item{dens}{- (list) output from characteristic_function_to_density}
}
\value{
vec - list [[1]] dim one expectiation, variance, skewness, kurtosis
                    [[2]] dim two expectiation, variance, skewness, kurtosis
                    [[3]] covariance
}
\description{
takes output from characteristic_function_to_density and computes
mean, variance, skewness, kurtosis
}
