% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charfuncUtil.R
\name{calc_moment_emperical}
\alias{calc_moment_emperical}
\title{takes sample vector and computes
mean, variance, skewness, kurtosis}
\usage{
calc_moment_emperical(X, central = F)
}
\arguments{
\item{X}{- (n x 1) output from iid sample of densit}
}
\value{
vec - (4x1) emperical expectiation, variance, skewness, kurtosis
}
\description{
takes sample vector and computes
mean, variance, skewness, kurtosis
}
