% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charfuncUtil.R
\name{characteristic_function_to_density}
\alias{characteristic_function_to_density}
\title{log density from char func using fft}
\usage{
characteristic_function_to_density(logphi, n2, interv, logphi_prev = NULL)
}
\arguments{
\item{logphi}{- (function) should return expontial term of the char func}

\item{n2}{- (1x1) power of 2 for number of point evaluations}

\item{interv}{- (2x1) evalution interval}
}
\description{
log density from char func using fft
}
