% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charfuncUtil.R
\name{characteristic_function_to_density2d}
\alias{characteristic_function_to_density2d}
\title{log density from char func using fft \eqn{\phi(x,y)}
on the grid [a_x,b_x] x [a_y, b_y]}
\usage{
characteristic_function_to_density2d(
  logphi,
  n2_x,
  n2_y,
  interv_x,
  interv_y,
  logphi_prev = NULL
)
}
\arguments{
\item{logphi}{- (function) logphi(x,y) should return \eqn{log( \phi(x,y))}
the grid of kron(x,1) and kron(y,1)}

\item{n2_x}{- (1) power of 2 for number of point evaluations of x dir}

\item{n2_y}{- (1) power of 2 for number of point evaluations of y dir}

\item{interv_x}{- (2x1) [a_x,b_x]}

\item{interv_y}{- (2x1) [a_y,b_y]}
}
\description{
log density from char func using fft \eqn{\phi(x,y)}
on the grid [a_x,b_x] x [a_y, b_y]
}
