% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.generation.R
\name{create.meshes.1d}
\alias{create.meshes.1d}
\title{Create meshes.}
\usage{
create.meshes.1d(locs, n, common.grid, extend = NULL)
}
\arguments{
\item{locs}{A numeric list for the meansurement locations.}

\item{common.grid}{A logical variable for using a common grid for all subjects,
\code{"TRUE"} indicates using a common grid,
\code{"FALSE"} uncommon grids.}

\item{extend}{A numeric vector with two elements specifying the amount of extension
of the grid to the left and right beyondthe measurement locations.}
}
\value{
Returns a list of meshes.
}
\description{
A function to create mesh for FEM discretization.
}
\details{
This is a supplementary function to be used internally by other functions.
}
\examples{
  \dontrun{
  create_meshes.1d(...)
  }

}
\seealso{
\code{\link{create_matrices_Matern}}, \code{\link{create_matrices_FD2}}
}
