% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{create_matrices_Matern}
\alias{create_matrices_Matern}
\title{creates matrices for Matern 1D operator}
\usage{
create_matrices_Matern(locs, n, right.boundary = "neumann",
  left.boundary = "neumann", common.grid = TRUE, extend = NULL)
}
\value{
operator_List list to to use in simulation and estimation object
}
\description{
creates matrices for Matern 1D operator
}
