% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{create_operator_matern2D}
\alias{create_operator_matern2D}
\title{Compute FEM matrices - 2D.}
\usage{
create_operator_matern2D(mesh)
}
\arguments{
\item{mesh}{An \code{inla.mesh} object.}
}
\description{
A function to compute FEM matrices for 2D (spatial) problems.
}
\details{
This is a supplementary function to be used internally by other functions.
}
\examples{
  \dontrun{
  create_operator_matern2D(...)
  }

}
