% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{create_operator_matern2Dbivariate}
\alias{create_operator_matern2Dbivariate}
\title{Compute FEM matrices - 2D bivariate fields.}
\usage{
create_operator_matern2Dbivariate(mesh, n.rep = 1)
}
\arguments{
\item{mesh}{An \code{inla.mesh} object.}
}
\description{
A function to compute FEM matrices for 2D (spatial) bivariate problems.
}
\details{
This is a supplementary function to be used internally by other functions.
}
\examples{
  \dontrun{
  create_operator_matern2Dbivariate(...)
  }

}
