% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharFunc.R
\name{density_1d_normal}
\alias{density_1d_normal}
\title{computes density for normal univarate random fields in 1d
between [interv[0], interv[1]]}
\usage{
density_1d_normal(interv, n2, param, maternParam)
}
\arguments{
\item{interv}{- start and endpoint to evalute density}

\item{n2}{- number of grid points in the interval}

\item{param}{- (2 x 1)  mu, sigma}

\item{maternParam}{- (3 x 1)  nu, tau, kappa}
}
\description{
computes density for normal univarate random fields in 1d
between [interv[0], interv[1]]
}
