% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharFunc.R
\name{density_2d_nig_multivariate}
\alias{density_2d_nig_multivariate}
\title{Density for bivariate NIG random fields on a square in R2}
\usage{
density_2d_nig_multivariate(
  interv_x,
  interv_y,
  n2_x,
  n2_y,
  param_x,
  param_y,
  maternParam
)
}
\arguments{
\item{interv_x}{- (2x1) [a_x,b_x]}

\item{interv_y}{- (2x1) [a_y,b_y]}

\item{n2_x}{- (1) power of 2 for number of point evaluations of x dir}

\item{n2_y}{- (1) power of 2 for number of point evaluations of y dir}

\item{param_x}{- 
- (1) delta - location   parameter
- (2) mu    - assymteric parameter
- (3) sigma - scale      parameter
- (4) nu    - shape      parameter}

\item{param_y}{- 
- (1) delta - location   parameter
- (2) mu    - assymteric parameter
- (3) sigma - scale      parameter
- (4) nu    - shape      parameter}

\item{maternParam}{- list
[[1]] (3 x 1) -   nu_x, tau_x, kappa_x
[[2]] (3 x 1) -   nu_y, tau_y, kappa_y
[[3]] (2)    -   rho, theta}
}
\description{
Computes the density for a random field on \eqn{[a_x,b_x] x  [a_y,b_y]}
}
