% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharFunc.R
\name{density_2d_normal}
\alias{density_2d_normal}
\title{Density for bivariate Gaussian stochastic process on an interval
computes density for normal univarate random fields in 2d}
\usage{
density_2d_normal(interv, n2, param, maternParam)
}
\arguments{
\item{interv}{- start and endpoint to evalute density}

\item{n2}{- number of grid points in the interval}

\item{param}{- (2 x 1)  mu, sigma}

\item{maternParam}{- (3 x 1)  nu, tau, kappa}
}
\description{
Density for bivariate Gaussian stochastic process on an interval
computes density for normal univarate random fields in 2d
}
