% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharFunc.R
\name{density_2d_normal_multivariate}
\alias{density_2d_normal_multivariate}
\title{Density for bivariate Gaussian random fields on a square in R2}
\usage{
density_2d_normal_multivariate(
  interv_x,
  interv_y,
  n2_x,
  n2_y,
  param_x,
  param_y,
  maternParam
)
}
\arguments{
\item{interv_x}{- (2x1) [a_x,b_x]}

\item{interv_y}{- (2x1) [a_y,b_y]}

\item{n2_x}{- (1) power of 2 for number of point evaluations of x dir}

\item{n2_y}{- (1) power of 2 for number of point evaluations of y dir}

\item{param_x}{- (2 x 1)  mu, sigma}

\item{param_y}{- (2 x 1)  mu, sigma}

\item{maternParam}{- list
[[1]] (3 x 1) -   nu_x, tau_x, kappa_x
[[2]] (3 x 1) -   nu_y, tau_y, kappa_y
[[3]] (2)    -   rho, theta}
}
\description{
Density for bivariate Gaussian random fields on a square in R2
}
