% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dGH}
\alias{dGH}
\title{Density for Generalized hyperbolic distribution}
\usage{
dGH(x, delta, mu, sigma, a, b, p, logd = TRUE)
}
\arguments{
\item{x}{vector of points to be evaluated}

\item{delta}{location}

\item{mu}{assymetric}

\item{sigma}{sscale}

\item{a}{distribution parameter}

\item{b}{distribution parameter}

\item{p}{distribution parameter}
}
\description{
A function for evaluating the GH density.
}
