% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dig}
\alias{dig}
\title{Density for inverse Gaussian distribution}
\usage{
dig(x, a, b, log_in = F)
}
\arguments{
\item{x}{A numeric vector for quantiles.}

\item{a}{parameter}

\item{b}{parameter}
}
\description{
Density for inverse Gaussian distribution
}
