% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dnig}
\alias{dnig}
\title{Density function of Normal inverse Gaussian distribution.}
\usage{
dnig(x, delta, mu, nu, sigma)
}
\arguments{
\item{x}{A numeric vector for quantiles.}

\item{delta}{A numeric value for the location parameter.}

\item{mu}{A numeric value for the shift parameter.}

\item{nu}{A numeric value for the shape parameter.}

\item{sigma}{A numeric value for the scaling parameter.}
}
\value{
A list of outputs.
}
\description{
A function to calculate the value of the probability density 
   function of Normal inverse Gaussian distribution.
}
\details{
\eqn{f(x|\delta, \mu, \nu, \sigma) =} STUFF
}
\examples{
  \dontrun{
  dnig(...)
  }
}
