% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dnigMeasure}
\alias{dnigMeasure}
\title{Computing the density of the random scattered measure
defined by lesbeuge measure mA}
\usage{
dnigMeasure(x, mA, delta, mu, nu, sigma, log = F)
}
\description{
Computing the density of the random scattered measure
defined by lesbeuge measure mA
}
