% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHmixedInit.R
\name{estiamteMixedGH}
\alias{estiamteMixedGH}
\title{estimation function}
\usage{
estiamteMixedGH(input_list)
}
\arguments{
\item{input_list}{-}

\item{Y}{- list with the observations}

\item{Niter}{- number of iteration of the stochastic gradient}

\item{meas_noise}{- the aviable noise classes: Normal or NIG}

\item{noise}{- the distribution of the mixed effect}

\item{B_random}{- list for the random effect covariates (needs to be matrix, can be NULL)}

\item{B_fixed}{- list for the fixed  effect covariates (needs to be matrix, can be NULL)}

\item{beta_random}{- initial parameters of the random effect (mean parameter) (if not specified set to zero)}

\item{beta_fixed}{- initial parameters of the fixed  effect (if not specified set to zero)}

\item{Sigma}{- initial parameters of the covariance of random effect (if not specified set to I )}

\item{nu}{- shape parameter for noise (NIG only)}

\item{mu}{- shift parameter for noise (NIG only)}

\item{U}{- (list) inital guess of the random effect}

\item{V}{- (list) inital guess of the variance effect}

\item{meas_list}{- list for measurement error:}

\item{sigma}{- measurement noise variance}

\item{nu}{- shape parameter for noise (NIG only)}

\item{Vs}{- (list) inital guess for the noise of the measurement
output::}

\item{output_list}{- list}

\item{Y_list}{- the observations}

\item{mu}{-  tracjorty of the shift parameter (of the mixed effect)}

\item{nuVec}{-  tracjorty of the shape parameter (of the mixed effect)}

\item{nu_measerror}{-  tracjorty of the shape parameter (of the measurement error)}

\item{betaVec}{-  tracjorty of the random effect mean parameter}

\item{sigma_eps}{-  tracjorty of the standard devation parameter (of measurement error)}

\item{mixedeffect}{- list the output from mixed effects}

\item{noise}{- string of distribution type}

\item{B_random}{- list for the random effect covariates}

\item{B_fixed}{- list for the fixed  effect covariates}

\item{beta_random}{- the last iteration of beta_random (mean of random effect)}

\item{beta_fixed}{- the last iteration of beta_fixed}

\item{Sigma}{- the last iteration of Sigma ("covaraince of random effect")}

\item{nu}{- the last iteration of shape parameter}

\item{mu}{- the last iteration of shift parameter}

\item{U}{- the last sample of random effect}

\item{V}{- the last sample of the variance coeffient of random effect}

\item{measerror}{- list the output from measurement error}

\item{noise}{- string of distribution type of the measurement error}

\item{sigma}{- the last iteration of standard devation parameter}

\item{nu}{- the last iteration of shape parameter}

\item{Vs}{- the last sample of the random variance of the measurement error}
}
\description{
estimation function
}
