% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGH.R
\name{estimate.wrapper}
\alias{estimate.wrapper}
\title{Wrapper for parameter estimation.}
\usage{
estimate.wrapper(Y, locs, B_random, B_fixed, use.process = TRUE,
  operator.type = "fd2", n.process = NULL,
  measurement.distribution = "Normal",
  random.effect.distribution = "Normal", process.distribution = "Normal",
  individual.sigma = FALSE, silent = FALSE, estimation.options = NULL,
  estimate_fisher = FALSE, ...)
}
\arguments{
\item{Y}{A numeric list that contains outcome values.}

\item{locs}{A numeric list that contains the timings at which the outcomes 
are collected.}

\item{B_random}{A numeric list of random effects covariate matrices.}

\item{B_fixed}{A numeric list of fixed effects covariate matrices.}

\item{operator.type}{A character string for specifying the operator type. 
Available options are: 
\code{"fd2"} for integrated Random-Walk (integrated Brownian Motion), 
and \code{"matern"} for Matern family.}

\item{n.process}{A numerical value for the number of basis functions to
approximate the stochastic process.}

\item{measurement.distribution}{A character string to specify the distribution 
of the error term. Available options are: \code{"Normal"} for Normal, \code{"NIG"} 
for Normal-inverse Gaussian, \code{"tdist"} for t-distribution.}

\item{random.effect.distribution}{A character string that indicates the distribution 
of the random effects. Available options are:  \code{"Normal"} for Normal,
and \code{"NIG"} for Normal-inverve Gaussian distributions.}

\item{process.distribution}{A character vector that indicates the distribution of 
the process. Available options are: 
\code{"Normal"} for Normal, \code{"NIG"} for Normal-inverse Gaussian,
\code{"GAL"} for generalised-asymmetric Laplace, and \code{"CH"} for Cauchy
distributions.}

\item{individual.sigma}{A logical variable for specifying patient-specific mixture
random variable for the error-term; \code{"FALSE"} indicates do not obtain,
\code{"TRUE"} obtain.}

\item{silent}{A logical value for printing the details of the iterations;
\code{"TRUE"} indicates do not print, \code{"FALSE"} indicates print.}

\item{estimation.options}{A list of control inputs. See \code{"estimation.controls"} 
for the \code{"nglda_est"} function.}

\item{estimate_fisher}{A logical variable for whether Fisher-Information matrix
to be obtained; \code{"FALSE"} indicates do not obtain, \code{"TRUE"} obtain.}

\item{...}{Additional arguments.}
}
\value{
A list of fitted results.
}
\description{
A wrapper function for parameter estimation.
}
\details{
This function is a wrapper function (wraps \code{"estimateLong"}) 
   for parameter estimation. It internally selects the initial values to 
   start the stochastic gradient algorithn. The function is not advised to 
   be used. It is indeed called within wrapped by \code{"nglda_est"} that is a 
   more user-friendly function for parameter estimation.
}
\examples{
  \dontrun{
  data(srft_data)
  estimate.wrapper(...)
  }
}
\seealso{
\code{\link{nglda_est}}, \code{\link{estimateLong}}
}
