% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGH.R
\name{estimateLong}
\alias{estimateLong}
\title{Estimate parameters.}
\usage{
estimateLong(Y, locs, mixedEffect_list, measurment_list, processes_list,
  operator_list, step0 = 0.3, alpha = 0.3, learning_rate = 0,
  nBurnin_learningrate = NULL, nBurnin_base = 0, pSubsample = 1,
  polyak_rate = -1, subsample.type = 1, nPar_burnin = 0,
  pSubsample2 = 0.3, nIter = 10, nSim = 1, nBurnin = 10,
  silent = FALSE, seed = NULL, standardize.mixedEffects = FALSE,
  estimate_fisher = FALSE)
}
\arguments{
\item{Y}{A numeric list that contains outcome values.}

\item{locs}{A numeric list that contains the timings at which the outcomes 
are collected.}

\item{mixedEffect_list}{A list of inputs for random effects. 
\itemize{
\item \code{noise}       the distribution of the mixed effect
\item \code{B_random}    list for the random effect covariates (needs to be matrix, can be NULL)
\item \code{B_fixed}     list for the fixed  effect covariates (needs to be matrix, can be NULL)
\item \code{beta_random} initial parameters of the random effect (mean parameter) (if not specified set to zero)
\item \code{beta_fixed}  initial parameters of the fixed  effect (if not specified set to zero)
\item \code{Sigma}       initial parameters of the covariance of random effect (if not specified set to I )
\item \code{nu}          shape parameter for noise (NIG only)
\item \code{mu}          shift parameter for noise (NIG only)
\item \code{U}           (list) inital guess of the random effect
\item \code{V}           (list) inital guess of the variance effect
}}

\item{measurment_list}{A list of inputs for measurement error.
\itemize{
\item \code{sigma} measurement noise variance
\item \code{nu}    shape parameter for noise (NIG only)
\item \code{Vs}    (list) inital guess for the noise of the measurement
}}

\item{processes_list}{A list of inputs for the process.
\itemize{
\item \code{noise} either Normal, NIG or GAL (change name to type rather then noise)
\item \code{nu}    shape parameter for NIG or GAL
\item \code{mu}    assymetric parameter for NIG or GAL
}}

\item{step0}{- stepsize for optimizer is step0 / i^alpha}

\item{alpha}{- stepsize for optimizer is step0 / i^alpha}

\item{learning_rate}{- parameter for sthocastic gradient}

\item{nBurnin_learningrate}{- don't start learning before}

\item{pSubsample}{- precentage of data used in each gradient subsampling}

\item{polyak_rate}{- taking moving average of parameters (-1 means inactive, 0 mean pure mean)}

\item{subsample.type}{- Type of subsampling: 0 - uniform without replacement
1 - sample size weighted
3 - weighted sampling by gradient size}

\item{nPar_burnin}{- use "M-step" updates until this iteration.}

\item{pSubsample2}{- precentage of data used in each gradient subsampling weighted by gradient}

\item{nIter}{- number of iteration of the stochastic gradient}

\item{nSim}{- number of samples of the gibbs sampler to estimate the gradient}

\item{silent}{- print iteration info}

\item{seed}{- (unsinged int) seed for debuging}

\item{estimate_fisher}{A logical variable for whether Fisher-Information matrix
to be obtained; \code{"FALSE"} indicates do not obtain, \code{"TRUE"} obtain.}
}
\description{
A function that estimates parameters by 
   calling the \code{"estimateLong_cpp()"} function.
}
