% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGH.R
\name{estimateME}
\alias{estimateME}
\title{Parameter estimation without process.}
\usage{
estimateME(Y, mixedEffect_list, measurment_list, step0 = 0.3, alpha = 0.3,
  learning_rate = 0, nBurnin_learningrate = NULL, pSubsample = 1,
  pSubsample2 = 0.3, polyak_rate = -1, subsample.type = 1,
  nPar_burnin = 0, nBurnin_base = 0, nIter = 10, nSim = 1,
  nBurnin = 10, silent = FALSE, seed = NULL, estimate_fisher = FALSE)
}
\arguments{
\item{Y}{A numeric list that contains outcome values.}

\item{mixedEffect_list}{A list of inputs for random effects. 
\itemize{
\item \code{noise} The distribution of the mixed effects.
\item \code{B_random} A list that contains the random effect 
   covariates (needs to be matrix, can be NULL).
\item \code{B_fixed} A list that contains the fixed effect 
   covariates (needs to be matrix, can be NULL).
\item \code{beta_random} Initial values for the parameters of the 
   random effects (mean parameter) (if not specified set to zero).
\item \code{beta_fixed} Initial  values for the parameters of the 
   fixed effects (if not specified set to zero).
\item \code{Sigma} Initial values for the parameters of the 
   variance-covariance matrix of the random effects 
   (if not specified set to I ).
\item \code{nu} Shape parameter for noise (NIG only)
\item \code{mu} Shift parameter for noise (NIG only)
\item \code{U} A list of inital values of the random effects.
\item \code{V} A list of inital values of the variance effects.
}}

\item{measurment_list}{A list of inputs for measurement error.
\itemize{
\item \code{sigma} Measurement noise variance parameter.
\item \code{nu}    Shape parameter for noise (NIG only).
\item \code{Vs}    A list of inital values for the noise of the measurement.
}}

\item{step0}{A numeric value for stepsize for the optimizer; step0 / i^alpha.}

\item{alpha}{A numeric value for stepsize for the optimizer; step0 / i^alpha.}

\item{learning_rate}{A numeric value for the parameter of stochastic gradient.}

\item{nBurnin_learningrate}{A numeric value until which the learning will
not be started.}

\item{pSubsample}{A numeric value for the portion of data to be used in each
gradient iteration.}

\item{pSubsample2}{A numeric value for the portion of the data
to be used in each gradient subsampling weighted by gradient.}

\item{polyak_rate}{A numeric value for moving average of parameters;
-1: inactive, 0: pure mean.}

\item{subsample.type}{A numeric value for the type of subsampling;
1: uniform sampling, 
2: sample size weighted,
3: weighted sampling by gradient size, 
4: grouped sub-sampler.}

\item{nPar_burnin}{A numeric value; "M-step" updates will be used until this
iteration.}

\item{nBurnin_base}{A numerical value for burn-in simulations that are performed
for a subject that is sampled for the first time in the estimation method.}

\item{nIter}{A numeric value for the number of iteration that will be
used by the stochastic gradient.}

\item{nSim}{A numeric value for the number of samples of the Gibbs sampler
to estimate the gradient.}

\item{silent}{A logical value for printing the details of the iterations;
\code{"TRUE"} indicates do not print, \code{"FALSE"} indicates print.}

\item{seed}{A numerical value for starting the Gibbs samplers from fixed seed.}

\item{estimate_fisher}{A logical variable for whether Fisher-Information matrix
to be obtained; \code{"FALSE"} indicates do not obtain, \code{"TRUE"} obtain.}
}
\value{
A list of fitted results.
}
\description{
A function to estimate the model parameters without process component.
}
\details{
This function fits the classical Laird-Ware mixed effects model, i.e. 
   texcluding the process component, with Normal and more flexible distributions. 
   The function calls \code{"estimateLong_cpp()"} internally. 
   It is wrapped by \code{"estimate.wrapper"}, and is not advised to 
   be used.
}
\examples{
  \dontrun{
  data(srft_data)
  estimateME(...)
  }
}
\seealso{
\code{\link{nglda_est}}, \code{\link{estimate.wrapper}}
}
