% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.nglda_est.R
\name{fitted.nglda_est}
\alias{fitted.nglda_est}
\title{Fitted values.}
\usage{
\method{fitted}{nglda_est}(object, type = "marginal", ...)
}
\arguments{
\item{object}{A fitted object by calling \code{"summary.nglda_est"}.}

\item{type}{A character string for the type of fitted values;
\code{"marginal"} for fixed effects.}

\item{...}{Additional arguments; none used currently.}
}
\description{
Calculates fitted values.
}
\details{
Subject-specific fitted values can be obtained using \code{"nglda_predict"}.
}
\examples{
  \dontrun{
  fit <- nglda_est(...)
  fitted(fit)
  }

}
\seealso{
\code{\link{nglda_est}}, \code{\link{fitted}}
}
