% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.ngme.R
\name{fitted.ngme}
\alias{fitted.ngme}
\title{Fitted values.}
\usage{
\method{fitted}{ngme}(object, type = "marginal", ...)
}
\arguments{
\item{object}{A fitted object by calling \code{"ngme"} function.}

\item{type}{A character string for the type of fitted values;
\code{"marginal"} for fixed effects.}

\item{...}{Additional arguments; none used currently.}
}
\description{
Calculates fitted values.
}
\details{
Subject-specific fitted values can be obtained using \code{"predict.ngme"}.
}
\examples{
  \dontrun{
  fit <- ngme(...)
  fitted(fit)
  }

}
\seealso{
\code{\link{ngme}}, \code{\link{fitted}}
}
