% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.generation.R
\name{generate.1d.mesh}
\alias{generate.1d.mesh}
\title{Create 1d mesh from observation locations.}
\usage{
generate.1d.mesh(x, max.dist, cutoff = 1e-10, extend = NULL,
  Y = NULL, loc.Y = NULL, max.dY = -1, nJump = 3,
  fix.differences = TRUE)
}
\arguments{
\item{x}{A list of measurement locations.}

\item{max.dist}{The largest distance between nodes in the mesh}

\item{cutoff}{Merge nodes in x that are closer than cutoff}
}
\value{
Returns a list with mesh nodes, distances, and number of nodes.
}
\description{
A function to create mesh for 1d FEM discretization.
}
\details{
This is a supplementary function to be used internally by other functions.
}
\examples{
  \dontrun{
  x = c(0,0.05,0.2,0.25,0.5,0.7,0.8,0.95,0.99,1)
  x <- 2*sort(runif(n))
  m <- generate.1d.mesh(x,max.dist = 0.2,cutoff = 0.1)
 plot(x,0*x)
 points(m$s,0*m$s,pch=4,col=2)
  }

}
\seealso{
\code{\link{create_matrices_Matern}}, \code{\link{create_matrices_FD2}}
}
