% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.generation.R
\name{generate.adaptive.meshes.1d}
\alias{generate.adaptive.meshes.1d}
\title{Create 1d mesh from observation locations adaptively.}
\usage{
generate.adaptive.meshes.1d(locs, max.dist = NULL, cutoff = 1e-10,
  common.grid = FALSE, extend = NULL, n.cores = 1)
}
\arguments{
\item{locs}{A list of measurement locations.}

\item{max.dist}{The largest distance between nodes in the mesh}

\item{cutoff}{Merge nodes in x that are closer than cutoff}

\item{common.grid}{A logical variable for using a common grid for all subjects,
\code{"TRUE"} indicates using a common grid,
\code{"FALSE"} uncommon grids.}

\item{extend}{A numeric vector with two elements specifying the amount of extension
of the grid to the left and right beyondthe measurement locations.}

\item{n.cores}{A numeric value for the number of cores to be used.}
}
\value{
Returns a list of output.
}
\description{
A function to create mesh for 1d FEM discretization adaptively.
}
\details{
This is a supplementary function to be used internally by other functions.
}
\examples{
  \dontrun{
  generate.adaptive.meshes.1d(...)
  }

}
\seealso{
\code{\link{create.meshes.1d}}
}
