% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{group.fixed}
\alias{group.fixed}
\title{Group individuals for grouped sub-sampler.}
\usage{
group.fixed(B)
}
\arguments{
\item{B}{A numeric list that contains fixed effects covariate matrices}
}
\value{
Returns a list of groups and a vector free.
}
\description{
A function to group individuals based on fixed effect matrix B.
}
\details{
This function is supplementary and internally used.
Each element in groups contains a vector of elements sufficient to make t(B)*B full rank.
}
\examples{
  \dontrun{
  scale.fixed(...)
  }

}
\seealso{
\code{\link{estimateLong}}
}
