% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharFunc.R
\name{logchar_dens_matern}
\alias{logchar_dens_matern}
\title{log charateristic function of \int f(x)dW_x assumes f is istorpic!
where W_x brownian sheet}
\usage{
logchar_dens_matern(t, logphi_W, maternParam, d, typeMatern = 1)
}
\description{
t           - (vector) location of chara eval
logphi_W    - (function)   evaluates log of char function of f(x)*W(dx)
maternParam - (3 x 1)  nu, tau, kappa
d           - (1 x 1)  dimension
}
