% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharFunc.R
\name{logchar_f_W}
\alias{logchar_f_W}
\title{exponent in characteristic function}
\usage{
logchar_f_W(u, logphi_W, f, range)
}
\arguments{
\item{u}{- point where to evalute \eqn{log(\phi(x))}}

\item{logphi_W}{- (function) evaluates log of Char function of f(x)*W(dx)}

\item{f}{- (function) the kernel}

\item{range}{- (2 x 1)    a,b}
}
\description{
Calculate exponent in characteristic function of \eqn{\int_{a}^b f(x) W(dx)}.
}
