% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharFunc.R
\name{logchar_nig_f_eval}
\alias{logchar_nig_f_eval}
\title{calculate exponent in characteristic function.}
\usage{
logchar_nig_f_eval(x, u, param, f)
}
\arguments{
\item{x}{-}

\item{u}{-}

\item{param}{(1) delta - location   parameter
- (2) mu    - assymteric parameter
- (3) sigma - scale      parameter
- (4) nu    - shaper     parameter}

\item{f}{- function kernel}
}
\description{
calculate exponent in characteristic function.
}
