% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharFunc.R
\name{logchar_nig_multi_f_eval}
\alias{logchar_nig_multi_f_eval}
\title{calculate exponent in characteristic function.}
\usage{
logchar_nig_multi_f_eval(x, u_x, u_y, param, f_x, f_y)
}
\arguments{
\item{x}{- (real) location}

\item{u_x}{- (real) char func value}

\item{u_y}{- (real) char func value}

\item{param}{(4 x 1)
- (1) delta - location   parameter
- (2) mu    - assymteric parameter
- (3) sigma - scale      parameter
- (4) nu    - shaper     parameter}

\item{f_x}{- function kernel}

\item{f_y}{- function kernel}
}
\description{
\eqn{[f_1(x) f_2(x)] dW(x)} where W is a brownian motion
}
