% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharFunc.R
\name{logchar_normal_f_eval}
\alias{logchar_normal_f_eval}
\title{calculate exponent in characteristic function for normal
f(x) dW(x) where W is a brownian motion}
\usage{
logchar_normal_f_eval(x, u, param, f)
}
\arguments{
\item{x}{- (real) location}

\item{u}{- (real) char func value}

\item{param}{(2 x 1) parameter values
 - (1) delta - location   parameter
-  (2) sigma - scale      parameter}

\item{f}{- function kernel}
}
\description{
calculate exponent in characteristic function for normal
f(x) dW(x) where W is a brownian motion
}
