% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharFunc.R
\name{materncorr}
\alias{materncorr}
\title{Matern correlation function}
\usage{
materncorr(x, alpha, kappa, d)
}
\arguments{
\item{x}{- (n x 1) distance}

\item{alpha}{- (>0)    smoothness}

\item{kappa}{- (>0)    range-scale parameter}

\item{d}{- (int>0) dimension of x}
}
\value{
M(x) - (n x 1) correlation for Matern
}
\description{
Matern correlation function
}
