% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGH.R
\name{ME.startvalues}
\alias{ME.startvalues}
\title{Obtain initials for random effects.}
\usage{
ME.startvalues(Y, mixedEffect_list)
}
\arguments{
\item{Y}{A numeric list that contains outcome values.}

\item{mixedEffect_list}{A list for random effects.}
}
\value{
Returns a list for the mixed effects.
    See e.g. \code{mixedEffect_list} in the \code{"EstimateLong"} function.
}
\description{
A function to obtain initial values for the random effects.
}
\details{
The function treats random effects as fixed effects to
   obtain start values using ordinary least squares (OLS).
}
\examples{
  \dontrun{
  data(srft_data)
  ME.startvalues(...)
  }

}
\seealso{
\code{\link{ngme}}
}
