% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHmixedInit.R
\name{MeasurementErrorInit}
\alias{MeasurementErrorInit}
\title{measurement error input
TODO: generlize the class so that one put in only theta
      and then theta arguments should be passed on depending on noise class}
\usage{
MeasurementErrorInit(Y, noise = c("Normal", "NIG"), sigma = 1, nu = 1)
}
\arguments{
\item{Y}{- list with the observations}

\item{noise}{- the aviable noise classes}

\item{sigma}{- measurement variance}

\item{nu}{- shape parameter for noise (NIG only)}
}
\description{
measurement error input
TODO: generlize the class so that one put in only theta
      and then theta arguments should be passed on depending on noise class
}
