% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHmixedInit.R
\name{MixedInit}
\alias{MixedInit}
\title{Mixed effect input
TODO: generlize the class so that one put in only theta
      and then theta arguments should be passed on depending on noise class}
\usage{
MixedInit(noise = c("Normal", "NIG"), B_fixed = NULL, B_random = NULL)
}
\arguments{
\item{noise}{- the aviable noise classes}

\item{B_fixed}{- fixed effect}

\item{B_random}{- random effect}
}
\description{
Mixed effect input
TODO: generlize the class so that one put in only theta
      and then theta arguments should be passed on depending on noise class
}
