% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charfuncUtil.R
\name{moment_transform}
\alias{moment_transform}
\title{computes mean, variance, skewness, kurtosis from the four moments}
\usage{
moment_transform(moment)
}
\description{
computes mean, variance, skewness, kurtosis from the four moments
}
