% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nglda_predict.R
\name{nglda_predict}
\alias{nglda_predict}
\title{Prediction.}
\usage{
nglda_predict(object, id, type = "Filter", quantiles = c(0.025, 0.975),
  control_list = list(return.samples = TRUE, predict.derivaties = NULL,
  excursions = NULL, crps = TRUE, crps.skip = 1, nSim = 10, nBurnin = 10, silent
  = FALSE))
}
\arguments{
\item{object}{A fitted object obtained by calling \code{"nglda_est"}.}

\item{id}{A numeric vector containing the ID's of the subjects for whom
predictions are to be obtained.}

\item{type}{A character string for the type of prediction: \code{"Filter"} for
filtering, \code{"Smoothing"} for smoothing.}

\item{quantiles}{A two-elemnent vector that contains the quantiles
of the predictions to be calculated.}

\item{control_list}{A list of control variables.
\itemize{
\item \code{"return.samples"} A logical variable for returning the 
  Monte Carlo samples used to compute the predictions; \code{"TRUE"} indicates 
  return, \code{"FALSE"} do not return.
\item \code{"predict.derivatives"} A logical variable (????) to obtain 
  predictions for the derivatives of the process
\item \code{"excursions"} A list of excursion probabilities to compute. 
  Each list should contain: 
  \itemize{
  \item \code{"type"} - type of excursion '>' or '<',
  \item \code{"level"} - level to compute excursion probability for,
  \item \code{"process"} - which expression for the model, 
  \eqn{x\alpha + dU + W + Z} with \eqn{x \alpha} being fixed effects, 
  \eqn{dU} random effects and \eqn{Z} noise, to compute the probability for.  
  \code{'X'} for \eqn{x\alpha + dU + W}, 
  \code{'W'} for \eqn{W}, 
  \code{'Y'} for \eqn{x\alpha + dU + W + Z}, 
  \code{'Xderivative'} for the first derivarive of \eqn{x\alpha + dU + W}, 
  and  
  \code{'Wderivative'} for the first derivariate of \eqn{W}.
  } 
\item \code{"crps"} A logical variable for calculating 
  continuous ranked probability score (CRPS); \code{"TRUE"} indicates 
  calculate, \code{"FALSE"} do not calculate.
\item \code{"crps.skip"} A numerical value, say a, that indicates every \emph{a}th 
  element of the sample to be used to compute the crps score.
\item \code{"nSim"} A numerical value for the number of samples for the Gibbs sampler 
  to obtain the predictions. 
\item \code{"nBurnin"} A numeric value for the number of samples that are 
  discarded as burn-in whilst calculating the predictions.
\item \code{silent} A logical value for printing the details;
    \code{"TRUE"} indicates do not print, \code{"FALSE"} indicates print.
}}
}
\value{
A list of output.
}
\description{
Obtains predicted values based on filtering and smoothing distributions.
}
\details{
This function is a wrapper function that calls 
   \code{"predictLong"} internally.
}
\examples{
  \dontrun{
  nglda_predict(...)
  }
}
\seealso{
\code{\link{nglda_est}}
}
