% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngme.fisher.R
\name{ngme.fisher}
\alias{ngme.fisher}
\title{Estimation of Fisher information matrix}
\usage{
ngme.fisher(fit, std.threshold = NULL, observed = TRUE,
  silent = FALSE, n.cores = 1, nSim = 1000, n.rep = 10,
  nIter = NULL, only.effects = TRUE, nBurnin = 100)
}
\arguments{
\item{observed}{If TRUE, the observed Fisher information matrix is estimated. Otherwise the ordinary Fisher-Information matrix is estimated.}

\item{nIter}{A numeric value for the number of iterations to be used to
obtain the Fisher-Information matrix.}

\item{controls}{A list of control variables.
\itemize{
   \item \code{nSim} A numeric value for the number of samples of the Gibbs sampler
     to estimate the gradient.
   \item \code{pSubsample} A numeric value for the portion of data to be used in each
     gradient iteration. \code{pSubsample = 1} indicates use of all subjects' data.
   \item \code{nSim.fisher} A numeric value for the number of samples of the Gibbs sampler
     to obtain the Fisher-Information matrix.
   \item \code{nBurnin.base} A numerical value for burn-in simulations that are performed
     for a subject that is sampled for the first time in the estimation method.
   \item \code{subsample.type} A numeric value for the type of subsampling;
     1: uniform sampling,
     2: sample size weighted,
     3: weighted sampling by gradient size,
     4: grouped sub-sampler.
   \item \code{pSubsample2} A numeric value for the portion of the data
     to be used in each gradient subsampling weighted by gradient.
}}
}
\value{
The ngme result object with the Fisher information added.
}
\description{
Estimates Fisher information matrix for ngme result.
}
\examples{
  \dontrun{
  data(srft_data)
  ngme(...)
  }
}
