% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngme.par.R
\name{ngme.par}
\alias{ngme.par}
\title{Parameter estimation.}
\usage{
ngme.par(
  n.cores = 4,
  std.lim = 0.1,
  trend.lim = 0.01,
  max.rep = 10,
  controls = NULL,
  controls.init = NULL,
  use.process = TRUE,
  nIter = 1000,
  timeVar = NULL,
  location.names = NULL,
  init.fit = NULL,
  silent = FALSE,
  plot.type = "All",
  save.tracks = FALSE,
  ...
)
}
\arguments{
\item{n.cores}{Number of cores, and the number of parallel chains, to use. Default is 4.}

\item{std.lim}{Parameter for first convergence criterium. The estimation is stopped when the estimated Monte Carlo 
standard deviation for each parameter is less than std.lim times the parameter value. Default is 0.1.}

\item{trend.lim}{Parameter for second convergence criterium. The estimation is stopped when the rate of change
per batch of nIter iterations is not significantly larger than trend.lim times the current parameter value. Default is 0.01.}

\item{max.rep}{The total number of iterations that is run is given by \code{nIter*max.rep}. Convergence is checked
after every nIter iterations, and max.rep thus sets how many batches of nIter iterations that should be run at most. 
Default is 10.}

\item{nIter}{The number of iterations per batch of runs. Default is 1000.}

\item{plot.type}{Set to "All" to get parameter trajectories of all estimated parameters. However, at most 16 parameters
are plotted at once. Set to "TRUE" or "Fixed" to get plots of only the fixed effects.}

\item{save.tracks}{Save the individual parameter tracks for the parallel runs? Default FALSE.}

\item{...}{Other parameter needed by \code{\link{ngme}}}
}
\value{
A list of outputs.
}
\description{
Maximum likelihood model estimation using parallel runs of 
  stochastic gradient estimation. See \code{\link{ngme}} and \code{\link{ngme.spatial}} for explanation of the model specification.
}
\details{
The function calls \code{\link{ngme}} or \code{\link{ngme.spatial}} internally. See these functions for further information on the actual
model specification. When plots of parameter trajectories are shown, the gray lines show the trajectories of the individual
runs and the black curve is the estimate obtained by averaging the individual trajectories. The green lines show approximate
95% confidence bands for the estimate.
}
\examples{
  \dontrun{
  data(srft_data)

  # transform pwl to decrese it correlation with bage
  # then center all the covariates for better convergence
  srft_data$pwl2 <- with(srft_data, pwl - bage/1.5)
  srft_data[, c("sex_cent", "bage_cent", "fu_cent", "pwl2_cent")] <-
    scale(srft_data[, c("sex", "bage", "fu", "pwl2")], scale = FALSE)

  # fit the model with normal assumption for random effects, process and error
  # covariance function is integrated random walk
  set.seed(123)
  fit <- ngme.par(n.cores = 5, std.lim = 100,max.rep = 20,
                  fixed = log(egfr) ~ sex_cent + bage_cent + fu_cent + pwl2_cent,
                  random = ~ 1|id,
                  data = srft_data,
                  reffects = "Normal",
                  process = c("Normal", "fd2"),
                  error = "Normal",
                  timeVar = "fu",
                  use.process = TRUE,
                  silent = FALSE,
                  mesh = list(cutoff = 1/365,max.dist = 1/12,extend = 0.01),
                  controls = list(pSubsample = 0.025,subsample.type = 1))
}
}
\seealso{
\code{\link{ngme}}, \code{\link{ngme.spatial}}
}
\author{
David Bolin \email{davidbolin@gmail.com}
}
