% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngme.R
\name{ngme}
\alias{ngme}
\title{Parameter estimation.}
\usage{
ngme(fixed, random, data, timeVar, reffects = "Normal",
  process = c("Normal", "fd2"), error = "Normal", use.process = TRUE,
  silent = TRUE, controls = list(learning.rate = 0, polyak_rate = 0.1,
  nBurnin = 100, nSim = 2, nIter.gauss = 1000, nIter = 10000, pSubsample = 0.1,
  nPar_burnin = 0, nIter.fisher = 1000, nSim.fisher = 1000, step0 = 0.3, alpha =
  0.3, nBurnin_learningrate = NULL, nBurnin_base = 0, subsample.type = 4,
  pSubsample2 = 0.3, seed = NULL, standardize.mixedEffects = FALSE,
  estimate.fisher = TRUE, individual.sigma = FALSE, n.process = NULL))
}
\arguments{
\item{fixed}{A two-sided formula to specify the fixed effects design matrix.}

\item{random}{A one-sided formula to specify the random effects design matrix.}

\item{data}{A data-frame from which the response and covariates to be
extracted.}

\item{reffects}{A character string that indicates the distribution of the
random effects. Available options are:  \code{"Normal"} for Normal,
and \code{"NIG"} for Normal-inverve Gaussian distributions.}

\item{process}{A character vector with two elements to specify 
the process. Whilst the first element is for the covariance structure, the
second element is for the process distribution. Available options for the
first are: \code{"fd2"} for integrated Random-Walk (integrated Brownian
Motion), \code{"matern"} for Matern family; for the second element are:
\code{"Normal"} for Normal, \code{"NIG"} for Normal-inverse Gaussian,
\code{"GAL"} for generalised-asymmetric Laplace, and \code{"CH"} for Cauchy
distributions.}

\item{error}{A character string to specify the distribution of the error term.
Available options are: \code{"Normal"} for Normal, \code{"NIG"} for 
Normal-inverse Gaussian, \code{"tdist"} for t-distribution.}

\item{use.process}{A logical variable for inclusion of the stochastic process in
the mixed model: \code{"TRUE"} indicates inclusion, \code{"FALSE"} exclusion.}

\item{silent}{A logical value for printing the details of the iterations;
\code{"TRUE"} indicates do not print, \code{"FALSE"} indicates print.}

\item{controls}{A list of control variables for parameter estimation.
\itemize{
   \item \code{learning.rate} A numeric value for the parameter of stochastic gradient.
   \item \code{polyak_rate} A numeric value for moving average of parameters;
     -1: inactive, 0: pure mean.
   \item \code{nBurnin} A numeric value for the number of steps before starting
     gradient estimation.
   \item \code{nSim} A numeric value for the number of samples of the Gibbs sampler
     to estimate the gradient.
   \item \code{nIter.gauss} A numerical value for the number of iterations to be used
     to obtain the initial values for models with non-Gaussian processes.
   \item \code{nIter} A numeric value for the number of iteration that will be
     used by the stochastic gradient.
   \item \code{pSubsample} A numeric value for the portion of data to be used in each
     gradient iteration. \code{pSubsample = 1} indicates use of all subjects' data.  
   \item \code{nPar_burnin} A numeric value; "M-step" updates will be used until this
     iteration.
   \item \code{nIter.fisher} A numeric value for the number of iterations to be used to
     obtain the Fisher-Information matrix.
   \item \code{nSim.fisher} A numeric value for the number of samples of the Gibbs sampler
     to obtain the Fisher-Information matrix.
   \item \code{step0} A numeric value for stepsize for the optimizer; step0 / i^alpha.
   \item \code{alpha} A numeric value for stepsize for the optimizer; step0 / i^alpha.
   \item \code{nBurnin_learningrate} A numeric value until which the learning will
     not be started.
   \item \code{nBurnin_base} STUFF.   
   \item \code{subsample.type} A numeric value for the type of subsampling;
     1: uniform sampling, 
     2: sample size weighted,
     3: weighted sampling by gradient size, 
     4: grouped sub-sampler.
   \item \code{pSubsample2} A numeric value for the portion of the data
     to be used in each gradient subsampling weighted by gradient.
   \item \code{seed} A numerical value for starting the Gibbs samplers from fixed seed.
   \item \code{standardize.mixedEffects} A logical variable for standardising the covariates;
     \code{"FALSE"} indicates no standardisation, \code{"TRUE"} standardisation.
   \item \code{estimate.fisher} A logical variable for whether Fisher-Information matrix
     to be obtained; \code{"FALSE"} indicates do not obtain, \code{"TRUE"} obtain.
   \item \code{individual.sigma} A logical variable for specifying patient-specific mixture
     random variable for the error-term; \code{"FALSE"} indicates do not obtain,
     \code{"TRUE"} obtain.
   \item \code{n.process} A numerical value for the number of basis functions to
     approximate the stochastic process.
}}

\item{timevar}{A character string that indicates the name of the time variable.}
}
\value{
A list of outputs.
}
\description{
Estimates model parameters using maximum likelihood implemented by a
  computationally efficient stochastic gradient algorithm.
}
\details{
This function is a user-friendly wrapper that calls the \code{estimateLong} function. 
    Generic functions \code{summary}, \code{print} and \code{plot} are available for the 
    output returned by the function \code{ngme}.
}
\examples{
  \dontrun{
  data(srft_data)
  ngme(...)
  }
}
