% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CharFunc.R
\name{ngme.spatial.moment}
\alias{ngme.spatial.moment}
\title{Create margninal moment and density for ngme.spatial object
returns density and moments of various components of the model
in the form of mean, variance, skewness and kuriotos}
\usage{
ngme.spatial.moment(obj, n2x = 7, n2y = 7, sdx = 6, sdy = 6)
}
\arguments{
\item{obj}{- ngme.spatial}

\item{n2x}{- number 2^n2x grid point in char func calc}

\item{n2y}{- number 2^n2y grid point in char func calc (multivariate only)}

\item{sdx}{- number of +- standard devations to calc the density around}

\item{sdy}{- number of +- standard devations to calc the density around}
}
\description{
Create margninal moment and density for ngme.spatial object
returns density and moments of various components of the model
in the form of mean, variance, skewness and kuriotos
}
