% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngme.spatial.R
\name{ngme.spatial}
\alias{ngme.spatial}
\title{Parameter estimation of non-Gaussian spatial models.}
\usage{
ngme.spatial(
  fixed,
  random = NULL,
  fixed2 = NULL,
  random2 = NULL,
  group.id = NULL,
  use.process = TRUE,
  reffects = "Normal",
  process = c("Normal", "matern"),
  error = "Normal",
  error_assymetric = FALSE,
  data,
  location.names = NULL,
  silent = TRUE,
  nIter = 1000,
  mesh = NULL,
  controls = list(learning.rate = 0.3, polyak.rate = -1, nBurnin = 100, nSim = 2,
    pSubsample = NULL, nPar.burnin = 0, step0 = 1, alpha = 0.3, nBurnin.learningrate =
    NULL, nBurnin.base = 0, subsample.type = 1, pSubsample2 = 0.3, individual.sigma =
    FALSE, iter.start = 0),
  controls.init = list(learning.rate.init = 0.9, polyak.rate.init = -1, nBurnin.init =
    100, nSim.init = 2, nIter.init = 1000, pSubsample.init = 0.1, nPar.burnin.init = 0,
    step0.init = 0.9, alpha.init = 0.6, nBurnin.learningrate.init = NULL,
    nBurnin.base.init = 0, subsample.type.init = 0, pSubsample2.init = 0.3,
    individual.sigma.init = FALSE),
  init.fit = NULL,
  debug = FALSE
)
}
\arguments{
\item{fixed}{A two-sided formula to specify the fixed effects design matrix.}

\item{random}{A one-sided formula to specify the random effects design matrix (if any).}

\item{fixed2}{A two-sided formula to specify the fixed effects design matrix for the second dimension
for bivariate models.}

\item{random2}{A one-sided formula to specify the random effects design matrix (if any) for the second dimension
for bivariate models.}

\item{use.process}{A logical variable for inclusion of the stochastic process in
the mixed model. Default is \code{"TRUE"}.}

\item{reffects}{A character string that indicates the distribution of the
random effects if present in the model. Available options are:
\code{"Normal"} for Normal distribution, and
\code{"NIG"} for Normal-inverse Gaussian.}

\item{process}{A character string specifying the distribution of the driving noise of the 
spatial process  Available options are
  first are:
  \code{"Normal"} for Normal distribution,
  \code{"NIG"} for Normal-inverse Gaussian,
  \code{"GAL"} for generalised-asymmetric Laplace, and
  \code{"CH"} for Cauchy.
  For multivariate models, only \code{"Normal"} and \code{"NIG"} are currently available.}

\item{error}{A character string to specify the distribution of the error term.
Available options are:
\code{"Normal"} for Normal distribution,
\code{"NIG"} for Normal-inverse Gaussian,
\code{"tdist"} for t.
For mulivariatae models, only \code{"Normal"} is currently available.}

\item{error_assymetric}{if true the non-Gaussian error is assymetric}

\item{data}{A data-frame from which the response and covariates to be extracted.}

\item{location.names}{A character vector with the names of the spatial coordinates.}

\item{silent}{A logical value for printing the details of the iterations;
\code{"TRUE"} indicates do not print, \code{"FALSE"} print.}

\item{nIter}{A numeric value for the number of iteration that will be
used by the stochastic gradient.}

\item{mesh}{A mesh object of class inla.mesh}

\item{controls}{A list of control variables for parameter estimation. See \code{ngme} for details.}

\item{controls.init}{A list of control variables to be used to fit the normal model
to get the initial values for fitting a model with at least one of random effects,
process and error being non-Gaussian. See \code{ngme} for details.}

\item{init.fit}{A fitted \code{ngme.spatial} object to be used as a starting value for estimation.}
}
\value{
A list of outputs.
}
\description{
Likelihood-based parameter estimation of spatial non-Gaussian models.
}
\details{
The model that is estimated is of the form

\deqn{Y_{ij} = B(s_i)beta + U(s_i)beta_j + X_j(s_i) + e_{ij}}
Here i denots the index of the spatial location for the measurement and j denotes the number of the 
replicate in case of repeated measurements. The common mean value \eqn{B(s_i)beta} is specified using 
fixed effects, and the random effect part \eqn{U(s_i)beta_j} specifies a mean value varying between replicates. 
The process \eqn{X_j(s)} is a spatial process (independent between replicates) with a Matern covariance 
structure, specified as a stochastic PDE with possibly non-Gaussian driving noise. 
The measurement noise \eqn{e_{ij}} is assumed to be independent between observations and replicates.
}
\seealso{
\code{\link{predict.ngme.spatial}}
}
