% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGH.R
\name{operator.startvalues.bivariate}
\alias{operator.startvalues.bivariate}
\title{Obtain initial values for the operator.}
\usage{
operator.startvalues.bivariate(
  Y,
  locs,
  mixedEffect_list,
  operator_list,
  measurement_list
)
}
\arguments{
\item{Y}{A numeric list that contains outcome values.}

\item{locs}{A numeric list that contains the timings at which the outcomes
are collected.}

\item{mixedEffect_list}{A list for random effects.}

\item{operator_list}{A list for operator.}

\item{measurement_list}{A list for measurement error.}
}
\value{
A list for operator.
   See e.g. \code{"operator_list"} in \code{"estimateLong"}.
}
\description{
A function to obtain initials for the bivariate operator.
}
\examples{
  \dontrun{
  operator.startvalues(...)
  }

}
