% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orthodont.R
\docType{data}
\name{Orthodont}
\alias{Orthodont}
\title{Orthodont data}
\format{A data frame with 108 rows and 5 columns:
\describe{
\item{distance}{outcome variable}
\item{age}{age of the child}
\item{Subject}{id as character}
\item{Sex}{sex of the child}
\item{id}{id as numeric}
}}
\usage{
Orthodont
}
\description{
A data-set that contains dental data for 27 children at ages 8, 10, 12 and 14
}
\keyword{datasets}
