% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nglda_est.R
\name{plot.nglda_est}
\alias{plot.nglda_est}
\title{Trace plots.}
\usage{
\method{plot}{nglda_est}(object, param = "fixed", ...)
}
\arguments{
\item{object}{A fitted object by calling \code{"nglda_est"}.}

\item{param}{A character string for which parameter set the traces
are to be plotted;
\code{"fixed"} for fixed effects,
\code{"random"} for time-invariant random effects,
\code{"process"} for stochastic process,
\code{"error"} for error term.}

\item{...}{Additional arguments; none used currently.}
}
\description{
Plots the parameter estimates at successive iterations
   of the stochastic algorithm.
}
\examples{
  \dontrun{
  fit <- nglda_est(...)
  fitted(fit)
  }

}
\seealso{
\code{\link{nglda_est}}
}
