% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nglda_predict.R
\name{plot.nglda_predict}
\alias{plot.nglda_predict}
\title{Prediction plots.}
\usage{
\method{plot}{nglda_predict}(object, id, control = list(xlim_x = 0.1, ylim_c =
  0.01), ...)
}
\arguments{
\item{object}{A fitted object by calling \code{"nglda_est"}.}

\item{id}{A numerical value or character string for ID of the subject
for whom the plot will be generated.}

\item{control}{A list of control variables.
\itemize{
\item \code{"xlim_x"} A numerical value to control the range of x-axis.
\item \code{"ylim_c"} A numerical value to control the range of y-axis.
}}

\item{...}{Additional arguments; none used currently.}
}
\description{
Plots the predicted values for a specific subject.
}
\examples{
  \dontrun{
  pred <- nglda_predict(...)
  plot(fit, 1)
  }

}
\seealso{
\code{\link{nglda_predict}}
}
