% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ngme.R
\name{plot.ngme}
\alias{plot.ngme}
\title{Trace plots.}
\usage{
\method{plot}{ngme}(object, param = "fixed", n.exclude = 0, index = NULL,
  ...)
}
\arguments{
\item{object}{A fitted object returned by the \code{"ngme"} function.}

\item{param}{A character string for which parameter set the traces
are to be plotted;
\code{"fixed"} for fixed effects,
\code{"random"} for time-invariant random effects,
\code{"process"} for stochastic process,
\code{"error"} for error term.}

\item{n.exclude}{A numeric value for excluding a portion of the 
trajectory (from left) for plotting.}

\item{...}{Additional arguments; none used currently.}
}
\description{
Plots the parameter estimates at successive iterations
   of the stochastic algorithm.
}
\examples{
  \dontrun{
  fit <- ngme(...)
  fitted(fit)
  }

}
\seealso{
\code{\link{ngme}}
}
