% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.predict.ngme.R
\name{plot.predict.ngme}
\alias{plot.predict.ngme}
\title{Prediction plots.}
\usage{
\method{plot}{predict.ngme}(object, id, control = list(xlim_x = 0.1, ylim_c =
  0.01), ...)
}
\arguments{
\item{object}{A fitted object returned by the \code{"predict.ngme"} function.}

\item{id}{A numerical value or character string for ID of the subject
for whom the plot will be generated.}

\item{control}{A list of control variables.
\itemize{
\item \code{"xlim_x"} A numerical value to control the range of x-axis.
\item \code{"ylim_c"} A numerical value to control the range of y-axis.
}}

\item{...}{Additional arguments; none used currently.}
}
\description{
Plots the predicted values for a specific subject.
}
\examples{
  \dontrun{
  fit <- ngme(...)
  pred <- predict(fit, ...)
  plot(pred, 1)
  }

}
\seealso{
\code{\link{predict.ngme}}
}
