% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{plot_GH_noise}
\alias{plot_GH_noise}
\title{Plot for mixed effects model fit.}
\usage{
plot_GH_noise(res, dRE = c())
}
\arguments{
\item{res}{A fitted object from mixed model fit.}

\item{dRE}{A numeric vector for dimension to plot.}
}
\value{
A plot.
}
\description{
A function to plot the results of mixed effects model fit.
}
\details{
STUFF.
}
\examples{
  \dontrun{
  plot_GH_noise(...)
  }
}
