% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyak.ngme.R
\name{polyak.ngme}
\alias{polyak.ngme}
\title{Post-processing of ngme parameter estimates}
\usage{
polyak.ngme(object, param = "fixed", polyak.rate = 0, plot = TRUE,
  burnin.end = NULL)
}
\arguments{
\item{object}{A fitted object returned by the \code{"ngme"} function.}

\item{param}{A character string for which parameter set the averaging should be performed;
\code{"fixed"} for fixed effects,
\code{"random"} for time-invariant random effects,
\code{"process"} for stochastic process,
\code{"error"} for error term.}

\item{param}{plot Plot the results or not?}
}
\description{
Updates the results using Polyak averaging and estimates the MC variance of the parameter estimates.
}
\examples{
  \dontrun{
  fit <- ngme(...)
  fitted(fit)
  }

}
\seealso{
\code{\link{ngme}}
}
