% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ngme.spatial.R
\name{predict.ngme.spatial}
\alias{predict.ngme.spatial}
\title{Prediction.}
\usage{
\method{predict}{ngme.spatial}(
  object,
  id = NULL,
  type = "Smoothing",
  data = NULL,
  quantiles = c(0.025, 0.975),
  controls = list(return.samples = FALSE, excursions = NULL, crps = TRUE, nSim = 1000,
    nBurnin = 100, silent = TRUE, seed = NULL, n.cores = 1)
)
}
\arguments{
\item{object}{A fitted object obtained by calling \code{"ngme.spatial"}.}

\item{id}{A numeric vector containing the ID's of the replicates for whom
predictions are to be obtained. Default is set to \code{"NULL"}
indicating perform predictions for all replicates.}

\item{type}{A character string for the type of prediction: \code{"Smoothing"} gives spatial prediction
based on all available data, \code{"LOOCV"} gives leave-one-out crossvalidation where also a number of 
accuracy measures are calculated. For multivariate models, \code{"LOOCV"} gives cross-validation over 
observation locations (so the data for all dimensions are removed and predicted), whereas \code{"LOOCV"} 
gives crossvalidation over univariate observations.}

\item{quantiles}{A two-element vector that contains the quantiles
of the predictions to be calculated.}

\item{controls}{A list of control variables.
\itemize{
\item \code{"return.samples"} A logical variable for returning the
  Monte Carlo samples used to compute the predictions; \code{"TRUE"} indicates
  return, \code{"FALSE"} do not return.
\item \code{"excursions"} A list of excursion probabilities to compute.
   Each list should contain:
  \itemize{
  \item \code{"type"} Type of excursion for indicating calculation of the probabilities, 
   greater than or less than a threshol. User should specify "greater than" by \code{">"} 
   "less than" by \code{"<"}.
  \item \code{"level"} A numeric value for the threshold compute excursion probability for.
  \item \code{"process"} A character string that for which component of the model 
   the excursion probabilities to be calculated.
  \itemize{
  \item \code{'X'} for latent process plus fixed and random effects,
  \item \code{'W'} for latent process,
  \item \code{'Y'} for data,
  }
  }
\item \code{"crps"} A logical variable for calculating
  continuous ranked probability score (CRPS); \code{"TRUE"} indicates
  calculate, \code{"FALSE"} do not calculate.
\item \code{"nSim"} A numerical value for the number of samples for the Gibbs sampler
  to obtain the predictions.
\item \code{"nBurnin"} A numeric value for the number of samples that are
  discarded as burn-in whilst calculating the predictions.
\item \code{silent} A logical value for printing the details;
    \code{"TRUE"} indicates do not print, \code{"FALSE"} indicates print.
\item \code{n.cores} Number of cores to use for paralell computations, default is set to 1. If 
larger than one, predictions over different replicates are computed in paralell. 
}}
}
\value{
A list of output.
}
\description{
Spatial prediction and cross-validation.
}
\examples{
  \dontrun{
  fit <- ngme.spatial(...)
  predict(fit, ...)
  }
}
\seealso{
\code{\link{ngme.spatial}}
}
