% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predictLong}
\alias{predictLong}
\title{STUFF}
\usage{
predictLong(Y, locs, pInd, locs.pred, Brandom.pred = NULL, Bfixed.pred,
  return.samples = FALSE, type = "Filter", quantiles = NULL,
  predict.derivatives = NULL, excursions = NULL, crps = FALSE,
  crps.skip = 10, mixedEffect_list, measurment_list, processes_list,
  operator_list = NULL, nSim = 1, nBurnin = 10, silent = FALSE,
  max.num.threads = 2, repeat.mix = 10)
}
\arguments{
\item{pInd}{- indices of longitudinal samples to do prediction for}

\item{locs.pred}{- list with positions to predict}

\item{Brandom.pred}{- random effect covaraites at prediction locations}

\item{Bfixed.pred}{- fixed effect covaraites at prediction locations}

\item{return.samples}{- return samples used for prediction?}

\item{type}{- Type of prediction: Filter or Smoothing}

\item{quantiles}{- list of posterior quantiles to compute}

\item{excursions}{- list of excursion probabilities to compute. Each list should contain:
type  - type of excursion '>' or '<'.
level - level to compute excursion probability for
process - which process to compute the probability for, 'X', 'W', 'Y','Xderivative' or 'Wderivative'}
}
\description{
STUFF
}
