% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{print.intervals}
\alias{print.intervals}
\title{Print function for the function \code{intervals}.}
\usage{
\method{print}{intervals}(object, ...)
}
\arguments{
\item{object}{A fitted object by calling \code{"intervals"}.}

\item{...}{Additional arguments; none used currently.}
}
\description{
Prints the output produced by \code{intervals} function concisely.
}
\examples{
  \dontrun{
  fit <- nglda_est(...)
  intervals(fit)
  }
}
\seealso{
\code{\link{intervals}}
}
