% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nglda_est.R
\name{print.nglda_est}
\alias{print.nglda_est}
\title{Print function for \code{"nglda_est"} objects.}
\usage{
\method{print}{nglda_est}(object, ...)
}
\arguments{
\item{object}{A fitted object by calling \code{"nglda_est"}.}

\item{...}{Additional arguments; none used currently.}
}
\description{
Prints \code{"nglda_est"} objects concisely.
}
\examples{
  \dontrun{
  fit <- nglda_est(...)
  fit
  }
}
\seealso{
\code{\link{nglda_est}}, \code{\link{print}}
}
