% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nglda_predict.R
\name{print.nglda_predict}
\alias{print.nglda_predict}
\title{Print function for \code{"nglda_predict"} objects.}
\usage{
\method{print}{nglda_predict}(object, ...)
}
\arguments{
\item{object}{A fitted object by calling \code{"nglda_predict"}.}

\item{...}{Additional arguments; none used currently.}
}
\description{
Prints the output produced by \code{"nglda_predict"} concisely.
}
\details{
MAE, RMSE, CRPS, Interval coverage, Interval width
}
\examples{
  \dontrun{
  pred <- nglda_predict(...)
  pred
  }
}
\seealso{
\code{\link{nglda_predict}}, \code{\link{print}}
}
