% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ngme.R
\name{print.ngme}
\alias{print.ngme}
\title{Print function for \code{"ngme"} objects.}
\usage{
\method{print}{ngme}(object, ...)
}
\arguments{
\item{object}{A fitted object by calling \code{"ngme"}.}

\item{...}{Additional arguments; none used currently.}
}
\description{
A function to print results stored in the 
   objects returned by the \code{"ngme"} function.
}
\examples{
  \dontrun{
  fit <- ngme(...)
  fit
  }
}
\seealso{
\code{\link{ngme}}, \code{\link{print}}
}
