% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.functions.R
\name{print.predict.ngme}
\alias{print.predict.ngme}
\title{Print function for \code{"predict.ngme"} objects.}
\usage{
\method{print}{predict.ngme}(object, ...)
}
\arguments{
\item{object}{A fitted object by calling \code{"predict.ngme"}.}

\item{...}{Additional arguments; none used currently.}
}
\description{
Prints the output contained in the \code{"predict.ngme"} objects concisely.
}
\details{
MAE, RMSE, CRPS, Interval coverage, Interval width
}
\examples{
  \dontrun{
  fit <- ngme(...)
  pred <- predict.ngme(fit, ...)
  pred
  }
}
\seealso{
\code{\link{predict.ngme}}, \code{\link{print}}
}
