% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.nglda_est.R
\name{print.summary.nglda_est}
\alias{print.summary.nglda_est}
\title{Print function for \code{"summary.nglda_est"} function.}
\usage{
\method{print}{summary.nglda_est}(object, ...)
}
\arguments{
\item{object}{A fitted object by calling \code{"summary.nglda_est"}.}

\item{...}{Additional arguments; none used currently.}
}
\description{
Prints \code{"summary.nglda_est"} objects concisely.
}
\examples{
  \dontrun{
  fit <- nglda_est(...)
  summary(fit)
  }

}
\seealso{
\code{\link{nglda_est}}, \code{\link{summary.nglda_est}},
  \code{\link{print}}, \code{\link{summary}}
}
