% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.ngme.R
\name{print.summary.ngme}
\alias{print.summary.ngme}
\title{Print function for \code{"summary.ngme"} function.}
\usage{
\method{print}{summary.ngme}(object, ...)
}
\arguments{
\item{object}{A fitted object returned by the \code{"summary.ngme"} function.}

\item{...}{Additional arguments; none used currently.}
}
\description{
A function to print \code{"summary.ngme"} objects concisely.
}
\examples{
  \dontrun{
  fit <- ngme(...)
  summary(fit)
  }

}
\seealso{
\code{\link{ngme}}, \code{\link{summary.ngme}},
  \code{\link{print}}, \code{\link{summary}}
}
