% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.nglda_est.R
\name{residuals.nglda_est}
\alias{residuals.nglda_est}
\title{Residual values.}
\usage{
\method{residuals}{nglda_est}(object, type = "marginal", ...)
}
\arguments{
\item{object}{A fitted object by calling \code{"summary.nglda_est"}.}

\item{type}{A character string for the type of fitted values;
\code{"marginal"} for fixed effects.}

\item{...}{Additional arguments; none used currently.}
}
\description{
Calculates residual values.
}
\details{
Subject-specific residaul values are not available currently.
}
\examples{
  \dontrun{
  fit <- nglda_est(...)
  fitted(fit)
  }

}
\seealso{
\code{\link{nglda_est}}, \code{\link{fitted.nglda_est}}
}
