% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngme.utils.R
\name{residuals.ngme}
\alias{residuals.ngme}
\title{Residual values.}
\usage{
\method{residuals}{ngme}(object, type = "marginal", ...)
}
\arguments{
\item{object}{A fitted object returned by the \code{"summary.ngme"} function.}

\item{type}{A character string for the type of fitted values;
\code{"marginal"} for fixed effects.}

\item{...}{Additional arguments; none used currently.}
}
\description{
Calculates residuals.
}
\details{
Subject-specific residaul values are not available currently.
}
\examples{
  \dontrun{
  fit <- ngme(...)
  fitted(fit)
  }

}
\seealso{
\code{\link{ngme}}, \code{\link{fitted.ngme}}
}
