% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGIG.R
\name{rGIG}
\alias{rGIG}
\title{Simulate generalised inverse-Gaussian (GIG) random variables.}
\usage{
rGIG(p, a, b, seed = 0)
}
\arguments{
\item{p}{A numeric vector for p parameters.}

\item{a}{A numeric vector for a parameters.}

\item{b}{A numeric vector for b parameters.}

\item{seed}{A numeric value for setting the seed of random number generation.}
}
\value{
A list of outputs.
}
\description{
A function to simulate random numbers 
    from GIG distribution -- GIG(p, a, b).
}
\details{
This function is a user-friendly wrapper that calls 
    the \code{rGIG_cpp} function. WE NEED PDF OF GIG(p, a, b).
}
\examples{
  \dontrun{
  rGIG(...)
  }
}
