% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.functions.R
\name{rNIG}
\alias{rNIG}
\title{Simulates nig distribution}
\usage{
rNIG(n = 1, delta, mu, sigma, nu, h = NULL)
}
\arguments{
\item{n}{number of simulations (if h is not NULL then n is ignored)}

\item{delta}{(real) location parameter}

\item{mu}{(real) symmetric parameter}

\item{sigma}{(real) scale parameter}

\item{nu}{(real, >0) shape parameter}

\item{h}{(m x 1) (h = 1) discritization vector}
}
\description{
internal function for simulating nig iid random variables using
             V ~ IG(nu,h^2 nu)
             Z ~ N(0, 1)
             X ~ delta + (V - h) * mu + sqrt(V) * sigma  * Z
}
