% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{scale.beta}
\alias{scale.beta}
\title{Scale fixed effects coefficients.}
\usage{
\method{scale}{beta}(beta, B.list, inv = FALSE)
}
\arguments{
\item{beta}{A numerical matrix of fixed effects estimates (in original scale).}

\item{B.list}{A numerical list of re-scaled covariate matrices.}

\item{inc}{A logical variable;
\code{"TRUE"} indicates the use of inverse re-scaling
(meaningful to use if \code{"sigma"} is given in transformed scale),
\code{"FALSE"} no use of inverse re-scaling.}
}
\value{
A matrix of scaled fixed effects coefficients.
}
\description{
A function to scale the fixed effects coefficients.
}
\details{
This function is supplementary and internally used.
}
\examples{
  \dontrun{
  scale.beta(...)
  }

}
\seealso{
\code{\link{estimateLong}}
}
