% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{scale.sigma}
\alias{scale.sigma}
\title{Scale covariance matrix for standardized covariates}
\usage{
\method{scale}{sigma}(sigma, B.list, inv = FALSE)
}
\arguments{
\item{sigma}{- Covariance matrix in original scale}

\item{B.list}{- list of rescaled covariates.}

\item{inv}{- set to true if inverse rescaleing is wanted (input sigma is given in transformed scale)}
}
\description{
Scale covariance matrix for standardized covariates
}
