% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGH.R
\name{setseed_ME}
\alias{setseed_ME}
\title{STUFF.}
\usage{
setseed_ME(input, seed)
}
\arguments{
\item{input}{STUFF}

\item{seed}{STUFF}
}
\value{
STUFF
}
\description{
STUFF
}
\details{
STUFF
}
\examples{
  \dontrun{
  setseed_ME(...)
  }

}
\seealso{
\code{\link{ngld_est}}, \code{\link{estimateLong}}, \code{\link{estimateME}}
}
