% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.functions.R
\name{simulate.process}
\alias{simulate.process}
\title{Simulates prior model using processes and operator}
\usage{
\method{simulate}{process}(n, operator_list, process_list)
}
\arguments{
\item{n}{number of simulations}

\item{operator_list}{operator list}

\item{process_list}{procsess list}
}
\value{
A list of simulation
}
\description{
Simulates prior model using processes and operator
}
