% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateLong.R}
\alias{simulateLong.R}
\title{Simulating longitudal model}
\usage{
simulateLong.R(loc, theta, n = 100, B = NULL, noise = c("Normal", "NIG",
  "GAL"), mixedEffect = FALSE, Bmixed = NULL, mixedEffectList = NULL,
  operatorType = "Matern", boundary = NULL)
}
\arguments{
\item{theta}{lost with covariates mu, kappa, sigma_eps, sigma}

\item{B}{list of matrix with covariates}

\item{locs}{list of location of observations}
}
\description{
Simulating longitudal model
}
