% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulateLongPrior}
\alias{simulateLongPrior}
\title{simulate data from the prior model}
\usage{
simulateLongPrior(Y, locs, mixedEffect_list, measurment_list, processes_list,
  operator_list)
}
\arguments{
\item{Y}{only used to get size of objects}

\item{locs}{measurement locations}

\item{mixedEffect_list}{mixed effects list}

\item{operator_list}{operator list}

\item{measurement_list}{measurement error list}

\item{process_list}{process list}
}
\description{
simulate data from the prior model
}
