% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{spde.A}
\alias{spde.A}
\title{Computes observation matrix.}
\usage{
spde.A(loc, x, right.boundary = "neumann", left.boundary = "neumann")
}
\arguments{
\item{loc}{A numeric vector of measurement locations.}

\item{x}{A numeric vector of node locations for the basis functions.}

\item{right.boundary}{A character string denoting the boundary condition 
for the right boundary.}

\item{left.boundary}{A character string denoting the boundary condition 
for the left boundary.}
}
\value{
Returns an observation matrix.
}
\description{
A function to compute observation matrix for 1D problems.
}
\details{
This is a supplementary function and internally used.
}
\examples{
  \dontrun{
  spde.A(...)
  }
 
}
\seealso{
\code{\link{build.A.matrix}}
}
