% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{spde.basis}
\alias{spde.basis}
\title{Compute FEM matrices.}
\usage{
spde.basis(x, right.boundary = "neumann", left.boundary = "neumann",
  compute.Ce = FALSE)
}
\arguments{
\item{x}{A numeric vector containing node locations.}

\item{right.boundary}{A character string denoting the
boundary condition for the right boundary.}

\item{left.boundary}{A character string denoting the
boundary condition for the left boundary.}
}
\value{
Returns a list that contains mass matrix, stiffness matrix,
   and some other related quantities.
}
\description{
A function to compute FEM matrices for 1D (longitudinal) problems.
}
\details{
This is a supplementary function and internally used.
}
\examples{
  \dontrun{
  spde.basis(...)
  }

}
\seealso{
\code{\link{create_matrices_Matern}}
}
