% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{standardize.covariates}
\alias{standardize.covariates}
\title{standardize covariates to avoid numerical issues}
\usage{
standardize.covariates(B.list)
}
\arguments{
\item{B.list}{- list with matrices containing covariates.}
}
\description{
standardize covariates to avoid numerical issues
}
