% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charfuncUtil.R
\name{sum_moment}
\alias{sum_moment}
\title{computes the centralized moment of X+Y
 @param moment1 - (4 x 1) four centralized moments of X
 @param moment2 - (4 x 1) four centralized moments of Y}
\usage{
sum_moment(moment1, moment2)
}
\description{
computes the centralized moment of X+Y
 @param moment1 - (4 x 1) four centralized moments of X
 @param moment2 - (4 x 1) four centralized moments of Y
}
