% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nglda_est.R
\name{summary.nglda_est}
\alias{summary.nglda_est}
\title{Summary function for \code{"nglda_est"} objects.}
\usage{
\method{summary}{nglda_est}(object, ...)
}
\arguments{
\item{object}{A fitted object by calling \code{"nglda_est"}.}

\item{...}{Additional arguments; none used currently.}
}
\description{
Summarises the output of \code{"nglda_est"} objects.
}
\examples{
  \dontrun{
  fit <- nglda_est(...)
  summary(fit)
  }
}
\seealso{
\code{\link{nglda_est}}, \code{\link{print.summary.nglda_est}},
  \code{\link{print}}, \code{\link{summary}}
}
