% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{create_matrices_FD2}
\alias{create_matrices_FD2}
\title{creates matrices for Finite difference operator, one sided}
\usage{
create_matrices_FD2(locs, n, right.boundary = "neumann",
  left.boundary = "neumann", common.grid = TRUE, extend = NULL)
}
\description{
creates matrices for Finite difference operator, one sided
}
