% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{create_matrices_Matern}
\alias{create_matrices_Matern}
\title{Create matrices for Matern.}
\usage{
create_matrices_Matern(
  locs,
  n,
  right.boundary = "neumann",
  left.boundary = "neumann",
  common.grid = FALSE,
  extend = NULL,
  max.dist,
  cutoff = 1e-10,
  n.cores = 1,
  name = "matern",
  beta = 1
)
}
\arguments{
\item{locs}{A numeric list for the meansurement locations.}

\item{n}{A numeric value for the number of FEM basis functions that should be used.}

\item{right.boundary}{A character string denoting the boundary condition
for the right boundary.}

\item{left.boundary}{A character string denoting the boundary condition
for the left boundary.}

\item{common.grid}{A logical variable for using a common grid for all subjects,
\code{"TRUE"} indicates using a common grid,
\code{"FALSE"} uncommon grids.}

\item{extend}{A numeric vector with two elements specifying the amount of extension
of the grid to the left and right beyondthe measurement locations.}

\item{max.dist}{A numeric value for largest distance between nodes in the discretization}

\item{name}{A character string for the operator type,
possible options are \code{"matern"}, \code{"exponential"} and \code{"fd2"}.}
}
\value{
Returns matrices.
}
\description{
A function to create matrices for Matern 1D (longitudinal) operator.
}
\details{
This is a supplementary function to be used internally by other functions.
}
\examples{
  \dontrun{
  create_matrices_Matern(...)
  }

}
\seealso{
\code{\link{create_operator}}
}
