% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateGH.R
\name{estimateME}
\alias{estimateME}
\title{estimating mixed effect model}
\usage{
estimateME(Y, mixedEffect_list, measurment_list, step0 = 0.3, alpha = 0.3,
  learning_rate = 0, nBurnin_learningrate = NULL, pSubsample = 1,
  pSubsample2 = 0.3, polyak_rate = -1, subsample.type = 1,
  nPar_burnin = 0, nBurnin_base = 0, nIter = 10, nSim = 1,
  nBurnin = 10, silent = FALSE, seed = NULL, estimate_fisher = FALSE)
}
\arguments{
\item{Y}{- list with the observations}

\item{mixedEffect_list}{-}

\item{measurment_list}{- list for measurement error:}

\item{step0}{- stepsize for optimizer is step0 / i^alpha}

\item{alpha}{- stepsize for optimizer is step0 / i^alpha}

\item{learning_rate}{- parameter for sthocastic gradient}

\item{pSubsample}{- precentage of data used in each gradient subsampling}

\item{polyak_rate}{- taking moving average of parameters (-1 means inactive, 0 mean pure mean)}

\item{subsample.type}{- Type of subsampling: 0 - uniform without replacement
1 - sample size weighted}

\item{nIter}{- number of iteration of the stochastic gradient}

\item{nSim}{- number of samples of the gibbs sampler to estimate the gradient}

\item{seed}{- (unsinged int) seed for debuging}

\item{locs}{- list with position of the observations (Y)}

\item{noise}{- the distribution of the mixed effect}

\item{B_random}{- list for the random effect covariates (needs to be matrix, can be NULL)}

\item{B_fixed}{- list for the fixed  effect covariates (needs to be matrix, can be NULL)}

\item{beta_random}{- initial parameters of the random effect (mean parameter) (if not specified set to zero)}

\item{beta_fixed}{- initial parameters of the fixed  effect (if not specified set to zero)}

\item{Sigma}{- initial parameters of the covariance of random effect (if not specified set to I )}

\item{nu}{- shape parameter for noise (NIG only)}

\item{mu}{- shift parameter for noise (NIG only)}

\item{U}{- (list) inital guess of the random effect}

\item{V}{- (list) inital guess of the variance effect}

\item{sigma}{- measurement noise variance}

\item{nu}{- shape parameter for noise (NIG only)}

\item{Vs}{- (list) inital guess for the noise of the measurement}

\item{operator_list}{- list created using create_operator function!}
}
\description{
estimating mixed effect model
}
