% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{intervals}
\alias{intervals}
\title{Confidence intervals.}
\usage{
intervals(object, type = "fixed", level = 0.95)
}
\arguments{
\item{object}{A fitted object by calling \code{"ngme"}.}

\item{type}{A character string for the type of parameters;
\code{"fixed"} for fixed effects.}

\item{level}{A numerical value indicating the confidence level.}
}
\description{
Calculates confidence intervals for model parameters.
}
\details{
Parameter estimates for random-effects and noise parameters
   are not available currently.
}
\examples{
  \dontrun{
  fit <- ngme(...)
  intervals(fit)
  }
  @export

}
\seealso{
\code{\link{ngme}}
}
