% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngme.fisher.R
\name{ngme.fisher}
\alias{ngme.fisher}
\title{Estimation of Fisher information matrix}
\usage{
ngme.fisher(
  fit,
  std.threshold = NULL,
  observed = TRUE,
  silent = FALSE,
  n.cores = 1,
  nSim = 1000,
  n.rep = 10,
  nIter = NULL,
  only.effects = TRUE,
  nBurnin = 100
)
}
\arguments{
\item{fit}{An ngme object.}

\item{std.threshold}{A threshold for the MC standard deviation of the estimates. The estimation is run until
all diagonal elements F.i of the inverse Fisher information matrix satisfy std(F.i)*std.threshold < F.i}

\item{observed}{If TRUE, the observed Fisher information matrix is estimated. Otherwise the ordinary Fisher-Information matrix is estimated.}

\item{silent}{If TRUE, some diagnostic information is shown during estimation.}

\item{n.cores}{The number of cores to use for the estimation. n.cores estimates are compued in parallel.}

\item{nSim}{A numeric value for the number of samples of the Gibbs sampler that is used internally.}

\item{n.rep}{A numeric value for the numer of MC estimates the initial estimate should be based on. If std.threshold is used
further estimates are computed until the criteria are satisfied.}

\item{nIter}{A numeric value for the number of iterations to be used to obtain the Fisher-Information matrix. If the observed Fisher information matrix
is estimate, nIter should be 1 since no new data has to be simulated.}

\item{only.effects}{If TRUE, the criteria for std.threshold is only applied to the fixed effect estimates.}

\item{nBurnin}{The number of samples to discard as burnin in the Gibbs sampler.}
}
\value{
The ngme result object with the Fisher information added.
}
\description{
Estimates Fisher information matrix for ngme result.
}
\examples{
  \dontrun{
  data(srft_data)
  ngme(...)
  }
  @export
}
