% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngme_package.R
\docType{package}
\name{ngme_package}
\alias{ngme_package}
\alias{ngme_package-package}
\title{ngme: An R package for non-Gaussian longitudinal data analysis}
\description{
The ngme package provides functions for parameter estimation and prediction.
}
\section{Function list}{
 ...
}

