% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ngme.R
\name{predict.ngme}
\alias{predict.ngme}
\title{Prediction.}
\usage{
\method{predict}{ngme}(object, newdata = NULL, id = NULL,
  type = "Filter", quantiles = c(0.025, 0.975),
  controls = list(return.samples = FALSE, predict.derivatives = NULL,
  excursions = NULL, crps = TRUE, crps.skip = 1, nSim = 1000, nBurnin =
  100, silent = TRUE, n.cores = 1, batch.size = 100))
}
\arguments{
\item{object}{A fitted object obtained by calling \code{"ngme"}.}

\item{id}{A numeric vector containing the ID's of the subjects for whom
predictions are to be obtained. Default is set to \code{"NULL"}
indicating perform predictions for all the subjects.}

\item{type}{A character string for the type of prediction: \code{"Nowcast"} 
for nowcasting, \code{"Filter"} for
one step-ahead forecasting, \code{"Smoothing"} for smoothing.}

\item{quantiles}{A two-element vector that contains the quantiles
of the predictions to be calculated.}

\item{controls}{A list of control variables.
\itemize{
\item \code{"return.samples"} A logical variable for returning the
  Monte Carlo samples used to compute the predictions; \code{"TRUE"} indicates
  return, \code{"FALSE"} do not return.
\item \code{"predict.derivatives"} A list for calculating excursion probabilities
  \itemize{
  \item \code{Bfixed} A list of fixed effects covariate matrices at \eqn{t_{ij}} + \code{delta}.
  \item \code{Brandom} A list of random effects covariate matrices at \eqn{t_{ij}}+ \code{delta}.
  \item \code{delta} A numeric value indicating \eqn{\Delta}t for calculating the derivative numerically.
  }
\item \code{"excursions"} A list of excursion probabilities to compute.
   Each list should contain:
  \itemize{
  \item \code{"type"} Type of excursion for indicating calculation of the probabilities, 
   greater than or less than a threshol. User should specify "greater than" by \code{">"} 
   "less than" by \code{"<"}.
  \item \code{"level"} A numeric value for the threshold compute excursion probability for.
  \item \code{"process"} A character string that for which component of the model 
   the excursion probabilities to be calculated.
  \itemize{
  \item \eqn{x\alpha + dU + W + Z} with \eqn{x \alpha} being fixed effects,
  \item \eqn{dU} random effects and \eqn{Z} noise, to compute the probability for.
  \item \code{'X'} for \eqn{x\alpha + dU + W},
  \item \code{'W'} for \eqn{W},
  \item \code{'Y'} for \eqn{x\alpha + dU + W + Z},
  \item \code{'Xderivative'} for the first derivarive of \eqn{x\alpha + dU + W},
  and
  \item \code{'Wderivative'} for the first derivariate of \eqn{W}.
  }
  }
\item \code{"crps"} A logical variable for calculating
  continuous ranked probability score (CRPS); \code{"TRUE"} indicates
  calculate, \code{"FALSE"} do not calculate.
\item \code{"crps.skip"} A numerical value, say a, that indicates every \emph{a}th
  element of the sample to be used to compute the crps score.
\item \code{"nSim"} A numerical value for the number of samples for the Gibbs sampler
  to obtain the predictions.
\item \code{"nBurnin"} A numeric value for the number of samples that are
  discarded as burn-in whilst calculating the predictions.
\item \code{silent} A logical value for printing the details;
    \code{"TRUE"} indicates do not print, \code{"FALSE"} indicates print.
\item \code{n.cores} Number of cores to use for paralell computations, default is set to 1.
\item \code{batch.size} Number of subjects to include in each paralell batch.
}}
}
\value{
A list of output.
}
\description{
Obtains predicted values based on filtering and smoothing distributions.
}
\details{
This function is a wrapper function that calls
   \code{"predictLong"} internally.
}
\examples{
  \dontrun{
  fit <- ngme(...)
  predict(fit, ...)
  }
}
\seealso{
\code{\link{ngme}}
}
