% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{scale.sigma}
\alias{scale.sigma}
\title{Scale covariance matrix.}
\usage{
\method{scale}{sigma}(sigma, B.list, inv = FALSE)
}
\arguments{
\item{sigma}{A numerical matrix of covariance matrix (in original scale).}

\item{B.list}{A numerical list of re-scaled covariate matrices.}
}
\value{
A list of scaled covariance matrix.
}
\description{
A function to scale covariance matrix.
}
\details{
This function is supplementary and internally used.
}
\examples{
  \dontrun{
  scale.sigma(...)
  }

}
\seealso{
\code{\link{estimateLong}}
}
