% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{standardize.covariates}
\alias{standardize.covariates}
\title{Standardise covariates.}
\usage{
standardize.covariates(B.list)
}
\arguments{
\item{B.list}{A numerical list that contains covariate matrices.}
}
\value{
A list of standardised covariate matrices.
}
\description{
A function to standardise covariates to avoid numerical issues.
}
\details{
This function is supplementary and internally used.
}
\examples{
  \dontrun{
  standardize.covariates(...)
  }

}
\seealso{
\code{\link{estimateLong}}
}
